+++ packages/sdk/package.json
@@ -36,6 +36,7 @@
     "dependencies": {
         "@dojoengine/torii-client": "workspace:*",
         "axios": "^0.27.2",
+        "immer": "^10.1.1",
         "lodash": "^4.17.21",
         "vite-plugin-wasm": "^3.3.0",
         "zustand": "^4.5.5"
